+++ internal/daemon/controller/gateway.go
@@ -29,7 +29,12 @@ import (
 	"google.golang.org/grpc/test/bufconn"
 )
 
-const gatewayTarget = ""
+const (
+	gatewayTarget = ""
+
+	// userAgentsKey defines the gRPC metadata key used to forward the User-Agent header to the gRPC server.
+	userAgentsKey = "userAgents"
+)
 
 type grpcServerListener interface {
 	net.Listener
@@ -58,6 +63,7 @@ func (noDelimiterStreamingMarshaler) Delimiter() []byte {
 func newGrpcGatewayMux() *runtime.ServeMux {
 	return runtime.NewServeMux(
 		runtime.WithMetadata(correlationIdAnnotator),
+		runtime.WithMetadata(userAgentHeadersAnnotator),
 		runtime.WithMarshalerOption(runtime.MIMEWildcard, &noDelimiterStreamingMarshaler{
 			&runtime.HTTPBodyMarshaler{
 				Marshaler: handlers.JSONMarshaler(),
@@ -90,6 +96,16 @@ func correlationIdAnnotator(_ context.Context, req *http.Request) metadata.MD {
 
 	return metadata.New(map[string]string{
 		globals.CorrelationIdKey: correlationId,
+	})
+}
+
+func userAgentHeadersAnnotator(_ context.Context, req *http.Request) metadata.MD {
+	userAgent := req.Header.Get("User-Agent")
+	if userAgent == "" {
+		return metadata.MD{}
+	}
+	return metadata.New(map[string]string{
+		userAgentsKey: userAgent,
 	})
 }