+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Canonical reference for changes, improvements, and bugfixes for Boundary.
 
 * Adds support for Azure Virtual Machine Scale Sets in the Azure plugin
   ([PR](https://github.com/hashicorp/boundary-plugin-azure/pull/22)).
+* Adds support to parse User-Agent headers and emit them in telemetry events
+  ([PR](https://github.com/hashicorp/boundary/pull/5645)).
 
 ### Bug fixes